+++ src/main/java/com/example/shop/service/ProductService.java
@@ -11,7 +11,7 @@ import java.util.Map;
 public interface ProductService {
     Product create(Product product) throws ResourceAlreadyExistsException;
     List<Product> getAll();
-    Product getOne(int id) throws ResourceNotFoundException;
-    Product remove(int id) throws ResourceNotFoundException;
-    Product update(int id, Map<String, Object> data) throws ResourceNotFoundException, ConstraintViolationException, Exception;
+    Product getOne(long id) throws ResourceNotFoundException;
+    Product remove(long id) throws ResourceNotFoundException;
+    Product update(long id, Map<String, Object> data) throws ResourceNotFoundException, ConstraintViolationException, Exception;
 }
